+++ src/main/java/edu/kis/vh/nursery/FirstInFirstOutRhymer.java
@@ -1,0 +1,20 @@
+package edu.kis.vh.nursery;
+
+public class FirstInFirstOutRhymer extends DefaultCountingOutRhymer {
+
+	public DefaultCountingOutRhymer temp = new DefaultCountingOutRhymer();
+
+	@Override
+	public int countOut() {
+		while (!callCheck())
+		    temp.countIn(super.countOut());
+
+		int ret = temp.countOut();
+
+		while (!temp.callCheck())
+		    countIn(temp.countOut());
+
+		return ret;
+	}
+}
+//Kombinacja ctrl+alt+ strzałka w lewo sprawia, że przechodzimy wstecz po histori aktualnie otwartych kart, z kolei ctrl + alt + strzałka w prawo idzie do przodu w kolejości otwartych kart.
