+++ src/main/java/edu/kis/vh/nursery/factory/RhymersFactory.java
@@ -1,0 +1,15 @@
+package edu.kis.vh.nursery.factory;
+
+import edu.kis.vh.nursery.DefaultCountingOutRhymer;
+
+	public interface RhymersFactory {
+	
+		public DefaultCountingOutRhymer GetStandardRhymer();
+		
+		public DefaultCountingOutRhymer GetFalseRhymer();
+		
+		public DefaultCountingOutRhymer GetFIFORhymer();
+		
+		public DefaultCountingOutRhymer GetHanoiRhymer();
+		
+	}
