--- src/main/java/edu/kis/vh/nursery/factory/Rhymersfactory.java
@@ -1,15 +1,0 @@
-package edu.kis.vh.nursery.factory;
-
-import edu.kis.vh.nursery.defaultCountingOutRhymer;
-
-	public interface Rhymersfactory {
-	
-		public defaultCountingOutRhymer GetStandardRhymer();
-		
-		public defaultCountingOutRhymer GetFalseRhymer();
-		
-		public defaultCountingOutRhymer GetFIFORhymer();
-		
-		public defaultCountingOutRhymer GetHanoiRhymer();
-		
-	}
